+++ .github/workflows/pr-instructions.yml
@@ -2,15 +2,32 @@ name: Add Pull Request Instructions
 
 on:
   pull_request:
-    types: [opened, edited]
+    types: [opened]
 
 jobs:
   Add-Pull-Request-Instructions:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - uses: actions/github-script@v4
+
+      # Create the message to post
+      - name: Create Message
+        uses: actions/github-script@v4
+        id: instruction
         with:
           script: |
-            const script = require('./github-actions/pr-instructions/pr-instructions.js')
-            console.log(await script({g: github, c: context}))
+            const script = require('./github-actions/pr-instructions/create-instruction.js')
+            const instruction = script({g: github, c: context})
+            return JSON.stringify({ instruction: instruction, issueNum: context.payload.number })
+      
+      # Create an artifact with the message
+      - name: Create Artifacts
+        run: |
+          mkdir -p addingPrInstructions/artifact
+          echo ${{ steps.instruction.outputs.result }} > addingPrInstructions/artifact/artifact.txt
+      
+      - name: Upload Artifacts
+        uses: actions/upload-artifact@v2
+        with:
+          name: adding-pr-instructions-artifact
+          path: addingPrInstructions/artifact/
